+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Added queue id constants
 * Added missing classes for game invitations
 * Added missing game modes
+* Added queue type and queue id for the new custom game mode, nemesis pick
 
 ##### Rtmp
 * Fixed bug where bot game invites would not pass the queue id